+++ 03-image-stats/c/ppm_stats.c
@@ -40,7 +40,7 @@ static int skip_field(FILE* f, int* line_num, int allow_hash) {
 // read_field: пропускает разделители, читает дес€тичное число.
 // ¬озвращает: 0 при успехе (число в *value), -1 при ошибке
 // (result->error и diagnostic заполнены)
-static int read_field(FILE* f, int* line_num, int allow_hash, long long* value,
+static int read_field(FILE* f, int* line_num, int allow_hash, int64_t* value,
                       struct StatsResult* result) {
     int c = skip_field(f, line_num, allow_hash);
     if (c == EOF) {
@@ -77,7 +77,7 @@ static int read_field(FILE* f, int* line_num, int allow_hash, long long* value,
         return -1;
     }
 
-    long long val = 0;
+    int64_t val = 0;
     while (c >= '0' && c <= '9') {
         val = val * 10 + (c - '0');
         if (val > 0x7FFFFFFF) {
@@ -99,7 +99,7 @@ static int read_field(FILE* f, int* line_num, int allow_hash, long long* value,
 
 // ”тилиты
 
-double luma(int r, int g, int b) { return 0.299 * r + 0.587 * g + 0.114 * b; }
+double luma(int32_t r, int32_t g, int32_t b) { return 0.299 * r + 0.587 * g + 0.114 * b; }
 
 // ѕубличные функции
 
@@ -155,7 +155,7 @@ struct StatsResult ppm_read_stats(FILE* f) {
     }
 
     // 2. Width
-    long long w_val = 0;
+    int64_t w_val = 0;
     if (read_field(f, &line_num, 1, &w_val, &result) != 0)
         return result;
     if (w_val <= 0 || w_val > 0x7FFFFFFF) {
@@ -165,10 +165,10 @@ struct StatsResult ppm_read_stats(FILE* f) {
                  w_val);
         return result;
     }
-    result.stats.width = (int)w_val;
+    result.stats.width = (int32_t)w_val;
 
     // 3. Height
-    long long h_val = 0;
+    int64_t h_val = 0;
     if (read_field(f, &line_num, 1, &h_val, &result) != 0)
         return result;
     if (h_val <= 0 || h_val > 0x7FFFFFFF) {
@@ -178,10 +178,10 @@ struct StatsResult ppm_read_stats(FILE* f) {
                  h_val);
         return result;
     }
-    result.stats.height = (int)h_val;
+    result.stats.height = (int32_t)h_val;
 
     // 4. Maxval
-    long long max_val = 0;
+    int64_t max_val = 0;
     if (read_field(f, &line_num, 1, &max_val, &result) != 0)
         return result;
     if (max_val != 255) {
@@ -191,15 +191,15 @@ struct StatsResult ppm_read_stats(FILE* f) {
                  line_num, max_val);
         return result;
     }
-    result.stats.max_val = (int)max_val;
+    result.stats.max_val = (int32_t)max_val;
 
     // 5. ѕопиксельные данные (PH_DATA Ч '#' не допускаетс€)
-    long long total_pixels = (long long)result.stats.width * result.stats.height;
-    int first_pixel = 1;
-    int pixel_error = 0;
+    int64_t total_pixels = (int64_t)result.stats.width * result.stats.height;
+    int32_t first_pixel = 1;
+    int32_t pixel_error = 0;
 
-    for (long long i = 0; i < total_pixels; ++i) {
-        long long r_val, g_val, b_val;
+    for (int64_t i = 0; i < total_pixels; ++i) {
+        int64_t r_val, g_val, b_val;
 
         if (read_field(f, &line_num, 0, &r_val, &result) != 0) {
             pixel_error = 1;
@@ -223,9 +223,9 @@ struct StatsResult ppm_read_stats(FILE* f) {
             return result;
         }
 
-        int r = (int)r_val;
-        int g = (int)g_val;
-        int b = (int)b_val;
+        int32_t r = (int32_t)r_val;
+        int32_t g = (int32_t)g_val;
+        int32_t b = (int32_t)b_val;
 
         result.stats.total_r += r;
         result.stats.total_g += g;
@@ -244,7 +244,7 @@ struct StatsResult ppm_read_stats(FILE* f) {
                 result.stats.y_max = y;
         }
 
-        int bin = (int)(y * 8.0 / (result.stats.max_val + 1));
+        int32_t bin = (int32_t)(y * 8.0 / (result.stats.max_val + 1));
         if (bin >= 8)
             bin = 7;
         result.stats.histogram[bin]++;
@@ -294,9 +294,9 @@ struct StatsResult ppm_read_stats(FILE* f) {
 }
 
 void ppm_print_stats(const struct Stats* s, FILE* out) {
-    int avg_r = (int)((double)s->total_r / s->pixel_count + 0.5);
-    int avg_g = (int)((double)s->total_g / s->pixel_count + 0.5);
-    int avg_b = (int)((double)s->total_b / s->pixel_count + 0.5);
+    int32_t avg_r = (int32_t)((double)s->total_r / s->pixel_count + 0.5);
+    int32_t avg_g = (int32_t)((double)s->total_g / s->pixel_count + 0.5);
+    int32_t avg_b = (int32_t)((double)s->total_b / s->pixel_count + 0.5);
 
     fprintf(out, "%dx%d\n", s->width, s->height);
     fprintf(out, "пикселей: %lld\n", s->pixel_count);
@@ -305,10 +305,10 @@ void ppm_print_stats(const struct Stats* s, FILE* out) {
     fprintf(out, "макс. €ркость: %.1f\n", s->y_max);
     fprintf(out, "гистограмма €ркости:\n");
 
-    for (int i = 0; i < 8; ++i) {
+    for (int32_t i = 0; i < 8; ++i) {
         double bin_width = (double)(s->max_val + 1) / 8.0;
-        int lo = (int)(i * bin_width);
-        int hi = (i == 7) ? s->max_val : (int)((i + 1) * bin_width) - 1;
+        int32_t lo = (int32_t)(i * bin_width);
+        int32_t hi = (i == 7) ? s->max_val : (int32_t)((i + 1) * bin_width) - 1;
         fprintf(out, "  [%3d,%3d]: %d\n", lo, hi, s->histogram[i]);
     }
 }
