+++ 03-image-stats/c/ppm_stats.h
@@ -1,9 +1,10 @@
 #ifndef KV_PPM_STATS_H
 #define KV_PPM_STATS_H
 
+#include <stdint.h>
 #include <stdio.h>
 
-#include "../../common/exit_codes.h"
+#include "../../common/c/exit_codes.h"
 
 #define STATS_DIAG_SIZE 256
 
@@ -19,16 +20,16 @@ enum StatsError {
 };
 
 struct Stats {
-    int width;
-    int height;
-    int max_val;
-    long long pixel_count;
-    long long total_r;
-    long long total_g;
-    long long total_b;
+    int32_t width;
+    int32_t height;
+    int32_t max_val;
+    int64_t pixel_count;
+    int64_t total_r;
+    int64_t total_g;
+    int64_t total_b;
     double y_min;
     double y_max;
-    int histogram[8];
+    int32_t histogram[8];
 };
 
 struct StatsResult {
